[refactor] Refactored routers
