+++ src/database/routers/utils.py
@@ -1,0 +1,9 @@
+from src.database.database import SessionLocal
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
